Separate S3Control methods from S3
